requirements.txt
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
transformers
pillow
fastapi
python-multipart
uvicorn
accelerate
safetensors